+++ Firmware/src/circular_buffer.h
@@ -12,16 +12,16 @@ typedef enum {
 
 typedef struct {
     uint8_t* data;
-    uint8_t head;
-    uint8_t tail;
-    uint8_t size;
+    uint16_t head;
+    uint16_t tail;
+    uint16_t size;
 } CircBuffer;
 
-void circBufferInit(CircBuffer* buffer, uint8_t* data, uint8_t size);
+void circBufferInit(CircBuffer* buffer, uint8_t* data, uint16_t size);
 bool isCircBufferEmpty(CircBuffer* buffer);
 BufferStatus circBufferPush(CircBuffer* buffer, uint8_t data);
 BufferStatus circBufferPop(CircBuffer* buffer, uint8_t* data);
-BufferStatus circBufferPopArray(CircBuffer* buffer, uint8_t* data, uint8_t length);
+BufferStatus circBufferPopArray(CircBuffer* buffer, uint8_t* data, uint16_t length);
 void circBufferClear(CircBuffer* buffer);
 
 #endif
